requirements.txt
Flask==2.2.5
requests
python-telegram-bot==20.3
gunicorn
